New attempt to fix the leaking companion problem
We now invalidate for all lifted symbols, whether they come form a tree or not.
